refactor: Update import paths for interfaces and adjust tsconfig settings
