+++ alx-project-0x03/pages/_app.tsx
@@ -1,6 +1,6 @@
 import type { AppProps } from 'next/app';
-import Layout from '../components/layouts/Layout'; // Adjusted import path
-import '../styles/globals.css'; // Adjusted import path
+import Layout from '../components/layouts/Layout';
+import '../styles/globals.css';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
