+++ alx-project-0x03/pages/index.tsx
@@ -1,11 +1,11 @@
-import Button from "../components/common/Button"; // Adjusted import path
-import { useRouter } from "next/router"; // For Pages Router
-import { PageRouteProps } from "../interfaces"; // Adjusted import path
+import Button from "../components/common/Button";
+import { useRouter } from "next/router";
+import { PageRouteProps } from "../interface";
 
 export default function Home() {
   const router = useRouter()
 
-  const routeToNextPage  = ({ pageRoute }: PageRouteProps) => {
+  const routeToNextPage = ({ pageRoute }: PageRouteProps) => {
     router.push(pageRoute);
   }
 
@@ -14,14 +14,15 @@ export default function Home() {
       <h1 className="text-4xl font-bold text-gray-800 mb-4">
         Welcome to Splash App!
       </h1>
+      <h1> Welcome to Splash App </h1>
       <p className="text-lg text-gray-600 mb-8">
         Your one-stop platform for everything AI you need. Start exploring by
         navigating to our features below.
       </p>
       <div className="flex gap-6">
         <Button action={() => routeToNextPage({ pageRoute: '/generate-text-ai' })} buttonLabel="Generate Text" buttonBackgroundColor="blue" />
-        <Button action={() => routeToNextPage({ pageRoute: '/text-to-image'})} buttonLabel="Text to Image" buttonBackgroundColor="green" />
-        <Button action={() => routeToNextPage({ pageRoute: '/counter-app'})} buttonLabel="Contact us" buttonBackgroundColor="orange" />
+        <Button action={() => routeToNextPage({ pageRoute: '/text-to-image' })} buttonLabel="Text to Image" buttonBackgroundColor="green" />
+        <Button action={() => routeToNextPage({ pageRoute: '/counter-app' })} buttonLabel="Contact us" buttonBackgroundColor="orange" />
       </div>
     </div>
   );
