+++ alx-project-0x03/tsconfig.json
@@ -3,7 +3,20 @@
     "jsx": "preserve",
     "esModuleInterop": true,
     "moduleResolution": "node",
-    "strict": false
+    "strict": false,
+    "target": "ES2017",
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
+    "allowJs": true,
+    "skipLibCheck": true,
+    "noEmit": true,
+    "incremental": true,
+    "module": "esnext",
+    "resolveJsonModule": true,
+    "isolatedModules": true
   },
   "include": [
     "next-env.d.ts",
